fix(tree): avoid sign comparison issues
stack-info: PR: https://github.com/aws/aws-ofi-nccl/pull/575, branch: aws-nslick/stack/22

+++ src/nccl_ofi_scheduler.c
@@ -58,7 +58,7 @@ void nccl_net_ofi_set_multiplexing_schedule(size_t size, int num_rails,
  */
 static inline int set_round_robin_schedule(nccl_net_ofi_threshold_scheduler_t *scheduler,
 					   size_t size,
-					   int num_rails,
+					   size_t num_rails,
 					   nccl_net_ofi_schedule_t *schedule)
 {
 	int rail_id;
